# specs/001-fullstack-todo-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo App Phase II API
  description: RESTful API for multi-user todo application with JWT authentication
  version: 1.0.0
  contact:
    name: Product Architect

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.example.com
    description: Production server

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for user
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create a new task
      operationId: createTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a specific task
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update a task
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion status
      operationId: toggleTaskComplete
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      description: User's unique identifier (must match authenticated user)
      schema:
        type: string
        format: uuid

    TaskId:
      name: task_id
      in: path
      required: true
      description: Task's unique identifier
      schema:
        type: string
        format: uuid

  schemas:
    Task:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
        user_id:
          type: string
          format: uuid
          description: Owner's user ID
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Task title
        description:
          type: string
          maxLength: 500
          nullable: true
          description: Optional task description
        is_completed:
          type: boolean
          description: Whether the task is completed
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Task title (required)
        description:
          type: string
          maxLength: 500
          nullable: true
          description: Optional task description

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Updated task title
        description:
          type: string
          maxLength: 500
          nullable: true
          description: Updated task description

    TaskList:
      type: object
      required:
        - tasks
        - count
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        count:
          type: integer
          description: Total number of tasks

    DeleteResponse:
      type: object
      required:
        - message
        - id
      properties:
        message:
          type: string
          example: Task deleted successfully
        id:
          type: string
          format: uuid
          description: ID of deleted task

    Error:
      type: object
      required:
        - detail
        - status
      properties:
        detail:
          type: string
          description: Error message
        code:
          type: string
          description: Error code for programmatic handling
        status:
          type: integer
          description: HTTP status code

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Validation error
            code: VALIDATION_ERROR
            status: 400

    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Invalid or missing authentication token
            code: UNAUTHORIZED
            status: 401

    Forbidden:
      description: User ID mismatch - accessing another user's resources
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: You do not have permission to access this resource
            code: FORBIDDEN
            status: 403

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: Task not found
            code: NOT_FOUND
            status: 404

tags:
  - name: Tasks
    description: Task management operations